requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23

# Data Analysis (updated for Python 3.14 compatibility)
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# Telegram Bot
python-telegram-bot>=20.7

# Scheduler
APScheduler>=3.10.4

# Utilities
pytz>=2023.3
python-dateutil>=2.8.2

# Development
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.1
flake8>=6.1.0
